requirements.txt
pandas==1.4.1
kafka-python==2.0.2
pymongo==4.1.0
requests==2.27.1